Fix keepAlive method in MqttService to convert KEEP_ALIVE_INTERVAL to Long correctly, ensuring proper MQTT connection behavior.

+++ app/src/main/java/com/example/cc/util/MqttService.kt
@@ -202,7 +202,7 @@ class MqttService : Service() {
             // Connect using HiveMQ client
             mqttClient.connectWith()
                 .cleanStart(true)
-                .keepAlive(MqttConfig.KEEP_ALIVE_INTERVAL.toLong())
+                .keepAlive(MqttConfig.KEEP_ALIVE_INTERVAL.toInt().toLong())
                 .send()
                 .whenComplete { connAck: Mqtt5ConnAck?, throwable: Throwable? ->
                     if (throwable != null) {
